+++ modules/common/src/types.ts
@@ -1,17 +1,17 @@
 export type Scorecard = {
-    courseName: string;
-    layoutName: string;
-    date: Date;
-    numberOfHoles?: number;
-    layoutPar?: number;
-    layoutDistance?: string;
-    entries: Array<ScorecardRecord>;
+  courseName: string;
+  layoutName: string;
+  date: Date;
+  numberOfHoles?: number;
+  layoutPar?: number;
+  layoutDistance?: string;
+  entries: Array<ScorecardRecord>;
 };
 
 export type ScorecardRecord = {
-    total: number;
-    players: Array<{
-      name: string;
-      username: string;
-    }>;
+  total: number;
+  players: Array<{
+    name: string;
+    username: string;
+  }>;
 };
